+++ hyswap/percentiles.py
@@ -618,7 +618,7 @@ def calculate_fixed_percentile_from_value(value, percentile_df):
         ...     data, 'values', percentiles=[5, 10, 25, 50, 75, 90, 95])
         >>> new_percentile = percentiles.calculate_fixed_percentile_from_value(
         ...     500, pcts_df)
-        >>> new_percentile
+        >>> new_percentile.astype(np.int64)
         50.0
 
     Calculate the percentiles associated with multiple values for some data
